move tests, fix some, change method call location
